refactor(runtime): isolate immutable desktop recovery plan
